Move base file class to util module.
