+++ tests/spells/nif/__init__.py
@@ -1,14 +1,14 @@
 import imp
 import os.path
 import sys
+import logging
 
 from os.path import dirname
 dir_path = __file__
 for i in range(4):  # recurse up to root repo dir
     dir_path = dirname(dir_path)
-repo_root = dir_path
 
-import logging
+repo_root = dir_path
 logger = logging.getLogger(__name__)
 
 logger.info(repo_root)
@@ -24,7 +24,3 @@ def call_niftoaster(*args):
     toaster.cli()
     sys.argv = oldargv
     return toaster
-
-
-class TestNif():
-    pass
